+++ processor/src/test/resources/input/com/example/dependency/CascadeDependencyModel.java
@@ -4,7 +4,7 @@ import arez.annotations.ArezComponent;
 import arez.annotations.Dependency;
 import arez.component.DisposeTrackable;
 
-@ArezComponent( allowEmpty = true )
+@ArezComponent
 public abstract class CascadeDependencyModel
 {
   @Dependency( action = Dependency.Action.CASCADE )