+++ processor/src/test/resources/input/com/example/dependency/ScheduleDeferredDependencyModel.java
@@ -4,8 +4,8 @@ import arez.annotations.ArezComponent;
 import arez.annotations.Dependency;
 import arez.component.DisposeTrackable;
 
-@ArezComponent( allowEmpty = true, deferSchedule = true )
-public abstract class ScheduleDeferredDependencyModel
+@ArezComponent( deferSchedule = true )
+abstract class ScheduleDeferredDependencyModel
 {
   @Dependency
   final DisposeTrackable getTime()