+++ processor/src/test/resources/input/com/example/dependency/TransitivelyDisposeTrackableDependencyModel.java
@@ -3,7 +3,7 @@ package com.example.dependency;
 import arez.annotations.ArezComponent;
 import arez.annotations.Dependency;
 
-@ArezComponent( allowEmpty = true )
+@ArezComponent
 public abstract class TransitivelyDisposeTrackableDependencyModel
 {
   @Dependency