+++ processor/src/test/resources/input/com/example/reference/LazyLoadObservableReferenceModel.java
@@ -6,7 +6,7 @@ import arez.annotations.Observable;
 import arez.annotations.Reference;
 import arez.annotations.ReferenceId;
 
-@ArezComponent( allowEmpty = true )
+@ArezComponent
 abstract class LazyLoadObservableReferenceModel
 {
   @Reference( load = LinkType.LAZY )