+++ processor/src/test/resources/input/com/example/reference/NullableLazyLoadReferenceModel.java
@@ -6,7 +6,7 @@ import arez.annotations.Reference;
 import arez.annotations.ReferenceId;
 import javax.annotation.Nullable;
 
-@ArezComponent( allowEmpty = true )
+@ArezComponent
 abstract class NullableLazyLoadReferenceModel
 {
   @Reference( load = LinkType.LAZY )